actualizacion 3.a

+++ src/com/humtrusa/entidades/Detalle_ventas.java
@@ -17,7 +17,7 @@ public class Detalle_ventas {
     private Long id_cabecera_venta;
     private Long id_producto;
     private String nombre_producto;
-    private int cantidad;
+    private Long cantidad;
     private BigDecimal precio;
     private BigDecimal subtotal;
     private BigDecimal iva;
@@ -27,7 +27,7 @@ public class Detalle_ventas {
     public Detalle_ventas() {
     }
 
-    public Detalle_ventas(Long id_detalle_venta, Long id_cabecera_venta, Long id_producto, String nombre_producto, int cantidad, BigDecimal precio, BigDecimal subtotal, BigDecimal iva, BigDecimal descuento, BigDecimal total) {
+    public Detalle_ventas(Long id_detalle_venta, Long id_cabecera_venta, Long id_producto, String nombre_producto, Long cantidad, BigDecimal precio, BigDecimal subtotal, BigDecimal iva, BigDecimal descuento, BigDecimal total) {
         this.id_detalle_venta = id_detalle_venta;
         this.id_cabecera_venta = id_cabecera_venta;
         this.id_producto = id_producto;
@@ -72,11 +72,11 @@ public class Detalle_ventas {
         this.nombre_producto = nombre_producto;
     }
 
-    public int getCantidad() {
+    public Long getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(Long cantidad) {
         this.cantidad = cantidad;
     }
 
@@ -119,5 +119,7 @@ public class Detalle_ventas {
     public void setTotal(BigDecimal total) {
         this.total = total;
     }
+
+
     
 }
